# _workspace/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PolitiKAST API",
    "description": "Korean local-election agent-based simulation backend.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/regions/{region_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Region",
        "operationId": "get_region_api_v1_regions__region_id__get",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions/{region_id}/summary": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Region Summary",
        "operationId": "get_region_summary_api_v1_regions__region_id__summary_get",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegionSummaryDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions/{region_id}/personas": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Sample Personas",
        "operationId": "sample_personas_api_v1_regions__region_id__personas_get",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Id"
            }
          },
          {
            "name": "n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 30,
              "title": "N"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonaSampleDTO"
                  },
                  "title": "Response Sample Personas Api V1 Regions  Region Id  Personas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions/{region_id}/poll-trajectory": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Region Poll Trajectory",
        "operationId": "region_poll_trajectory_api_v1_regions__region_id__poll_trajectory_get",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Id"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PollTrajectoryResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions/{region_id}/prediction-trajectory": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Region Prediction Trajectory",
        "operationId": "region_prediction_trajectory_api_v1_regions__region_id__prediction_trajectory_get",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Id"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionTrajectoryResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scenarios/{scenario_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Scenario",
        "operationId": "get_scenario_api_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scenarios/{scenario_id}/outcome": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Scenario Outcome",
        "operationId": "get_scenario_outcome_api_v1_scenarios__scenario_id__outcome_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOutcomeResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions/{region_id}/kg-subgraph": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Kg Subgraph",
        "operationId": "get_kg_subgraph_api_v1_regions__region_id__kg_subgraph_get",
        "parameters": [
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region Id"
            }
          },
          {
            "name": "cutoff",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO datetime cutoff",
              "title": "Cutoff"
            },
            "description": "ISO datetime cutoff"
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 25,
              "title": "K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KGSubgraphDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/anonymous": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Create Anonymous User",
        "operationId": "create_anonymous_user_api_v1_users_anonymous_post",
        "parameters": [
          {
            "name": "politikast_uid",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Politikast Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnonUserDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/comments/{comment_id}/report": {
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Report Comment",
        "operationId": "report_comment_api_v1_comments__comment_id__report_post",
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            }
          },
          {
            "name": "politikast_uid",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Politikast Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequestDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PollTrajectoryResponseDTO": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PollPointDTO"
            },
            "type": "array",
            "title": "Data"
          },
          "blackout": {
            "$ref": "#/components/schemas/BlackoutMetaDTO"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PollTrajectoryResponseDTO"
      },
...
      "RegionDTO": {
        "properties": {
          "region_id": {
            "type": "string",
            "title": "Region Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "election_id": {
            "type": "string",
            "title": "Election Id"
          },
          "election_date": {
            "type": "string",
            "title": "Election Date"
          },
          "position_type": {
            "type": "string",
            "title": "Position Type"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Asia/Seoul"
          },
          "in_blackout": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Blackout"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "region_id",
          "name",
          "election_id",
          "election_date",
          "position_type"
        ],
        "title": "RegionDTO"
      },
      "RegionSummaryDTO": {
        "properties": {
          "region_id": {
            "type": "string",
            "title": "Region Id"
          },
          "n_personas": {
            "type": "integer",
            "title": "N Personas",
            "default": 0
          },
          "n_districts": {
            "type": "integer",
            "title": "N Districts",
            "default": 0
          },
          "n_candidates": {
            "type": "integer",
            "title": "N Candidates",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "region_id"
        ],
        "title": "RegionSummaryDTO"
      },
...
    }
  }
}